formData Caption
I added those lines because i wanted to send only 1 file. But it still adds index number like 'file0'. But my backend server accept only 1 file. That's because if key 'file0' it thinks i am trying to send an array but not. It should be just 'file' if it is not multiple.